Reworked mutex code.
This was passing pthread_mutex_t around by value, but this is often a struct,
so we were copying the struct by-value, where we would lock the _copy_ that
is thrown away upon return from Mutex_Lock(), and the actual mutex will sit
there untouched, which is to say it basically disables locking. You might get
lucky and the struct has some internal pointer where you end up with a
working lock anyhow, but this is not the case on macOS and probably other
places, too.

Surprisingly, this very rarely produced a noticible race condition, but there
have been weird assertions from time to time, and I expect this will fix those.